Proposal to stop a possible panic:
In function `New(length uint) *BitSet`, it's possible to get a panic when giving it a value that is too big.

This a big problem for those of us who use the `bloom` package because the panic can occur when invalid files are passed in, or other manipulations where the cause -> effect isn't easily traceable. When I supply a file that doesn't work, I don't really know why it's giving me a panic on making a slice -- but I do understand that an unmarshalling error probably comes from an invalid data stream. It's also a good idea to remove any panic calls in a library, from what most documentation says.